refact(sharepoint-connector): moved steps in the processing-pipeline/steps directory
