+++ services/sharepoint-connector/src/processing-pipeline/processing-pipeline.module.ts
@@ -3,13 +3,13 @@ import { ConfigModule } from '@nestjs/config';
 import { AuthModule } from '../auth/auth.module';
 import { SharepointApiModule } from '../sharepoint-api/sharepoint-api.module';
 import { UniqueApiModule } from '../unique-api/unique-api.module';
-import { ContentFetchingStep } from './content-fetching.step';
-import { ContentRegistrationStep } from './content-registration.step';
+import { ContentFetchingStep } from './steps/content-fetching.step';
+import { ContentRegistrationStep } from './steps/content-registration.step';
 import { FileProcessingOrchestratorService } from './file-processing-orchestrator.service';
-import { IngestionFinalizationStep } from './ingestion-finalization.step';
+import { IngestionFinalizationStep } from './steps/ingestion-finalization.step';
 import { ProcessingPipelineService } from './processing-pipeline.service';
-import { StorageUploadStep } from './storage-upload.step';
-import { TokenValidationStep } from './token-validation.step';
+import { StorageUploadStep } from './steps/storage-upload.step';
+import { TokenValidationStep } from './steps/token-validation.step';
 
 @Module({
   imports: [ConfigModule, SharepointApiModule, AuthModule, UniqueApiModule],
