+++ services/sharepoint-connector/src/processing-pipeline/steps/content-registration.step.ts
@@ -1,11 +1,11 @@
 import { Injectable, Logger } from '@nestjs/common';
 import { ConfigService } from '@nestjs/config';
-import { UniqueAuthService } from '../auth/unique-auth.service';
-import { DEFAULT_MIME_TYPE } from '../constants/defaults.constants';
-import { UniqueApiService } from '../unique-api/unique-api.service';
-import type { IPipelineStep } from './steps/pipeline-step.interface';
-import type { ProcessingContext } from './types/processing-context';
-import { PipelineStep } from './types/processing-context';
+import { UniqueAuthService } from '../../auth/unique-auth.service';
+import { DEFAULT_MIME_TYPE } from '../../constants/defaults.constants';
+import { UniqueApiService } from '../../unique-api/unique-api.service';
+import type { IPipelineStep } from './pipeline-step.interface';
+import type { ProcessingContext } from '../types/processing-context';
+import { PipelineStep } from '../types/processing-context';
 
 @Injectable()
 export class ContentRegistrationStep implements IPipelineStep {
