+++ services/sharepoint-connector/src/processing-pipeline/steps/storage-upload.step.spec.ts
